💄 Extract Stadtteil
